Further updates

+++ chat_sdk_core/src/main/java/co/chatsdk/core/base/BaseContactHandler.java
@@ -1,10 +1,10 @@
 package co.chatsdk.core.base;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import co.chatsdk.core.NM;
-import co.chatsdk.core.dao.BUser;
-import co.chatsdk.core.dao.DaoCore;
+import co.chatsdk.core.dao.User;
 import co.chatsdk.core.handlers.ContactHandler;
 import co.chatsdk.core.types.ConnectionType;
 import io.reactivex.Completable;
@@ -16,24 +16,34 @@ import io.reactivex.Completable;
 public class BaseContactHandler implements ContactHandler {
 
     @Override
-    public List<BUser> contacts() {
-        return NM.currentUser().getContacts();
+    public List<User> contacts() {
+        if(NM.currentUser() != null) {
+            return NM.currentUser().getContacts();
+        }
+        return new ArrayList<>();
     }
 
     @Override
-    public List<BUser> contactsWithType(ConnectionType type) {
-        return NM.currentUser().getContacts(type);
+    public List<User> contactsWithType(ConnectionType type) {
+        if(NM.currentUser() != null) {
+            return NM.currentUser().getContacts(type);
+        }
+        return new ArrayList<>();
     }
 
     @Override
-    public Completable addContact(BUser user, ConnectionType type) {
-        NM.currentUser().addContact(user, type);
+    public Completable addContact(User user, ConnectionType type) {
+        if(NM.currentUser() != null) {
+            NM.currentUser().addContact(user, type);
+        }
         return Completable.complete();
     }
 
     @Override
-    public Completable deleteContact(BUser user, ConnectionType type) {
-        NM.currentUser().deleteContact(user, type);
+    public Completable deleteContact(User user, ConnectionType type) {
+        if(NM.currentUser() != null) {
+            NM.currentUser().deleteContact(user, type);
+        }
         return Completable.complete();
     }
 }
